% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Policy.R
\name{check}
\alias{check}
\title{Check that a policy allows access to a resource.}
\usage{
check(object, ...)
}
\arguments{
\item{object}{A policy to check}

\item{\dots}{method specific arguments}
}
\value{
Returns TRUE if policy allows access, FALSE otherwise
}
\description{
Check that a policy allows access to a resource.
}

