% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make-web-call.R
\name{format.content.type}
\alias{format.content.type}
\title{Performs the specific encoding required by the \code{Content-type} http
header.}
\usage{
\method{format}{content.type}(x, ...)
}
\arguments{
\item{x}{Description of content type as a list}

\item{...}{ignored}
}
\value{
A formatted \code{Content-type} string.
}
\description{
Implements the typical encoding for \code{Content-type}, which is semicolon
separated values or key-value pairs, where the pairs are slash-separated.
}
\details{
This function should only be used as a conversion argument when specifyng
the \code{Content-type} header.
}

