% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coor.R
\name{get_coor}
\alias{get_coor}
\title{Harvest coordinates using Google Geocode API}
\usage{
get_coor(Robj)
}
\arguments{
\item{Robj}{An object returned from the \code{\link{getURL2R}} function.}
}
\value{
A list containing the longitude, latitude and place
}
\description{
The \code{get_coor} function extracts the coordinates from an object returned from the \code{\link{getURL2R}} function.
}
\examples{
target<-getURL2R("Linkoping, Sweden")
CordObj<-get_coor(target)
}

