% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rka.R, R/get_rkaApp.R
\name{get_rka}
\alias{get_rka}
\title{To get the city information from RKA API}
\usage{
<<<<<<< HEAD
get_rka()
=======
get_rka(city, ageinterval)
>>>>>>> 93c54f73c10dda70c93907942e9a751057fcbde7
}
\arguments{
\item{city}{The name of the city in Sweden}

\item{ageinterval}{ageinterval of people}
}
\value{
a plot of the income county consule
}
\description{
To get the city information from RKA API

This is an example of the shiny app.
}
