% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURL2R.R
\name{getURL2R}
\alias{getURL2R}
\title{Harvest coordinates using Google Geocode API}
\usage{
getURL2R(path)
}
\arguments{
\item{path}{A character string of a street, city or an adress in the world}
}
\value{
A list containin the Geocode domain and the url used to get the coordinates.
}
\description{
The function gets the wanted coordinates using Google Geocode API
}
\examples{
getURL2R("Stockholm, Sweden")
getURL2R("58253, Linkoping")
}

