% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plot_map}
\alias{plot_map}
\title{Plotting the coordinates into a map}
\usage{
plot_map(dat, zoom = 13, typemap = "roadmap", ...)
}
\arguments{
\item{dat}{A object returned from the function  \code{\link{get_coor}}.}

\item{zoom}{A integer between 3 and 20 depending on how zoomed in you want the map, high values for big zoom.}

\item{typemap}{A character string of which kind of map you want. Possible choices are "roadmap","satellite","hybrid", "terrain-background" and "terrain". (See \code{\link{get_map}} in \code{ggmap} package)}
}
\value{
A map over the area from the coordinates in the object from \code{\link{get_coor}}
}
\description{
The function \code{plot_map} that plots the coordinates returned from the \code{\link{get_coor}} in a map of selectable types.
}
\examples{
target<-getURL2R("Linkoping, Sweden")
CordObj<-get_coor(target)
plot_map(CordObj, zoom=10, typemap="roadmap")
}

