% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToWebAPI.r
\name{urlConstructor}
\alias{urlConstructor}
\title{url Constructor}
\usage{
urlConstructor(l_kpi, l_m, l_year, url)
}
\arguments{
\item{l_kpi}{: Selected KPI codes}

\item{l_m}{: Selected Municipality Codes}

\item{l_year}{: Required Year}

\item{url}{: Required URL to retrieve the data from}
}
\value{
: returns the complete constructed URL
}
\description{
url Constructor
}
\details{
: This function will take the user input to construct an URL.
           Features: 1. User must input kpi code, municipality code and the year to fetch the data related to KPI vs Municipality for mentioned years.
                     2. Or, the User can give URL to fetch the data directly from the given URL.
           Note: if user inputs all the parameters then the data will be retrieved only from from the URL the user proveds.
}
\examples{

\dontrun{

urlConstructor("","","", url = "http://api.kolada.se/v2/municipality?all")
http://api.kolada.se/v2/municipality?all

urlConstructor(l_m = "0180,0581,0580",l_kpi = "N01951,N61714",l_year = "2018,2019,2020", url = "")
http://api.kolada.se/v2/data/kpi/N01951,N61714/municipality/0180,0581,0580/year/2018,2019,2020

#' urlConstructor(l_m = "0180,0581,0580",l_kpi = "N01951,N61714",l_year = "2018,2019,2020", url = "http://api.kolada.se/v2/municipality?all")
http://api.kolada.se/v2/municipality?all
}
}
