% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToWebAPI.r
\name{verifContent}
\alias{verifContent}
\title{Verify the Content}
\usage{
verifContent(response)
}
\arguments{
\item{response}{: Response from the getAPI function, this will contain the object of the API.}
}
\value{
: Json data level 1 flattened to Data.Frame
}
\description{
Verify the Content
}
\details{
: This function takes the response from the get API function, validates if the response has data content or not, finally parses the input json content and flatten the level 1 content.
}
