% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/webapi-package.R
\docType{package}
\name{webapi-package}
\alias{webapi}
\alias{webapi-package}
\title{Support for Creating an R API From a Web API.}
\description{
Create R functions from a description of each web API entry point.
}
\details{
Important functions:

\code{\link{make.web.call}}: Generate a R function call from a detailed description of a web API entry point

\code{\link{check}}: Check that a policy allows access to a resource

\code{\link{enforce}}: Enforce an access policy

\code{\link{update}}: Update a policy

Policy.rate.limit

arg.spec

format.content.type
}
\keyword{package}

