% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Policy.R
\name{enforce}
\alias{enforce}
\title{Enforce an access policy.}
\usage{
enforce(policy, ...)
}
\arguments{
\item{policy}{A policy to check}

\item{...}{method specific arguments}
}
\value{
Undefined
}
\description{
The rate limit policy enforces a delay necessary to comply with the rate limit
}

