% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rka.R
\name{get_rka}
\alias{get_rka}
\title{To get the city information from RKA API}
\arguments{
\item{city}{The name of the city in Sweden}
}
\value{
the informations of the city
}
\description{
To get the city information from RKA API
}
