% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToWebAPI.r
\name{getData}
\alias{getData}
\title{User function to get the data from API}
\usage{
getData(l_kpi = "", l_m = "", l_year = "", url = "")
}
\arguments{
\item{l_kpi}{: Required KPI Code}

\item{l_m}{: Required Municipality Code}

\item{l_year}{: Required Year}

\item{url}{: If data to be fetched from a specific URL}
}
\value{
: data.frame having the requested data from the API
}
\description{
User function to get the data from API
}
\details{
: This function to be called from the package to access the API, parse and download the data. The requested data will be downloaded as CSV file
}
\examples{
\dontrun{

final_data = getData(l_m = "0180,0581,0580",l_kpi = "N01951,N61714",l_year = "2018,2019,2020", url = "")
}

}
