% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURL2R.R
\name{getURL2R}
\alias{getURL2R}
\title{Get URL and harvest geodata using Google Geocode API}
\usage{
getURL2R(path)
}
\arguments{
\item{path}{A character string of postal number, names of the street, city or country in the world}
}
\value{
A list containing the Geodata about the address typed in the path such as coordinates.
}
\description{
The \code{getURL2R} takes the postal number, names of the street, city or country in the world encoded into a url. The url then returns a json formatted geodata and then converted to an R object.
}
\examples{
getURL2R("Stockholm, Sweden")
getURL2R("58253, Linkoping")
}

