% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Policy.R
\name{Policy.rate.limit}
\alias{Policy.rate.limit}
\title{Create a Policy.rate.limit object enforcing the namesake access policy}
\usage{
Policy.rate.limit(rate.limit, last.access = NULL)
}
\arguments{
\item{rate.limit}{The rate limit, in accesses per second.}

\item{last.access}{The date and time of the last access. This should always be left at default}
}
\value{
The newly minted Policy.rate.limit object
}
\description{
A rate limit policy prescribes a minimum interval between accesses to a resource
}
\details{
This is used to provide the \code{.policy} argument to \code{\link{make.web.call}}.
}

