% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Policy.R
\name{update}
\alias{update}
\alias{update.Policy}
\alias{update.Policy.rate.limit}
\title{Update a policy.}
\usage{
\method{update}{Policy}(object, ...)

\method{update}{Policy.rate.limit}(object, ...)
}
\arguments{
\item{object}{A policy to update}

\item{\dots}{Slots to update in the policy}
}
\value{
Undefined
}
\description{
A policy may contain internal state that needs to be updated (reference semantics). Use this method to do so
}
\details{
In the case of \code{\link{Policy.rate.limit}}, it is used internally by \code{\link{make.web.call}} to update the time of last access to a resource
}

