\name{webAPI-package}
\alias{webAPI-package}
\alias{webAPI}
\docType{package}
\title{
\packageTitle{webAPI}
}
\description{
\packageDescription{webAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{webAPI}
\packageIndices{webAPI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{webAPI}

Maintainer: \packageMaintainer{webAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
